melo_api/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==3.2.16
django-cloudinary-storage==0.3.0
django-countries==7.5
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
python-magic==0.4.27
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
